fiixed.c file

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -2,16 +2,16 @@
 #include "function_pointers.h"
 
 /**
- * array_iterator - executes a function given as a parameter on each element of an array.
+ * array_iterator - executes a function given as a parameter an array.
  * @array: pointer to array
- * set_t size: size of array
+ * @size: size of array
  * @action: pointer to function
  * Return: Nothing.
  */
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-long unsigned i;
+size_t i;
 if (array && action)
 {
 for (i = 0; i < size; i++)
